Implement tokenizer training and corpus tokenization functionality; remove obsolete tokenize.c
